imporving routing for apps, change edit form
