--- features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/source/src/app/components/Applications/Applications.jsx
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import React from 'react';
-import { Route, Switch } from 'react-router-dom';
-import ApplicationCreateForm from 'AppComponents/Applications/Create/NewApp';
-import Listing from './Listing/Listing';
-import Details from './Details/index';
-import { PageNotFound } from '../Base/Errors';
-/**
- *
- *
- * @class Applications
- * @extends {Component}
- */
-
-const applications = () => {
-    return (
-        <Switch>
-            <Route exact path='/applications' component={Listing} />
-            <Route exact path='/applications/create/' component={ApplicationCreateForm} />
-            <Route path='/applications/:application_uuid/' component={Details} />
-            <Route component={PageNotFound} />
-        </Switch>
-    );
-};
-
-
-export default applications;
