+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/source/src/app/components/Shared/AppsAndKeys/ApplicationCreateForm.jsx
@@ -191,7 +191,6 @@ const ApplicationCreate = (props) => {
             <TextField
                 margin='normal'
                 variant='outlined'
-                autoFocus
                 fullWidth
                 value={applicationRequest.description}
                 label={intl.formatMessage({
